src/use-cases/search-gyms.ts
import { Gym } from "@prisma/client";
import { GymsRepository } from "src/repositories/gyms-repository";

interface SearchGymsUseCaseRequest {
  query: string;
  page: number;
}

interface SearchGymsUseCaseResponse {
  gyms: Gym[];
}

export class SearchGymsUseCase {
  constructor(private gymsReposytory: GymsRepository) {}

  async execute({
    query,
    page,
  }: SearchGymsUseCaseRequest): Promise<SearchGymsUseCaseResponse> {
    const gyms = await this.gymsReposytory.searchMany(query, page);

    return {
      gyms,
    };
  }
}
